Tue, Sep 24, 2024, 9:14 PM +08:00

+++ readability/readability.c
@@ -51,7 +51,7 @@ int main(void){
         printf("Before Grade 1\n");
 
     else
-        printf("Grade %f\n", round(index));
+        printf("Grade %i\n", (int)round(index));
 
 
 }
